+++ sklearn_genetic/schedules/schedulers.py
@@ -2,9 +2,11 @@ import math
 from .base import BaseAdapter
 
 
-class ExponentialAdapter(BaseAdapter):
+class ConstantAdapter(BaseAdapter):
     """
-    Adapts the initial value towards the end value using an exponential "decay" function
+    This adapter keep the current value equals to the initial_value
+    it's mainly used to have an uniform interface when defining a
+    parameter as constant vs as an adapter
 
     Parameters
     ----------
@@ -21,7 +23,37 @@ class ExponentialAdapter(BaseAdapter):
             The current number of iterations that the adapter has run
 
         current_value : float,
-            The transformed initial_value after current_steps changes
+            Same as the initial_value
+    """
+
+    def __init__(self, initial_value, end_value, adaptive_rate):
+        super().__init__(initial_value, end_value, adaptive_rate)
+
+    def step(self):
+        self.current_step += 1
+        return self.current_value
+
+
+class ExponentialAdapter(BaseAdapter):
+    """
+    Adapts the initial value towards the end value using an exponential "decay" function
+
+    Parameters
+    ----------
+    initial_value : float,
+        Initial value to be adapted
+    end_value : float,
+        The final (asymptotic) value that the initial_value can take
+    adaptive_rate : float,
+        Controls how fast the initial_value approaches the end_value
+
+    Attributes
+    ----------
+    current_step : int,
+        The current number of iterations that the adapter has run
+
+    current_value : float,
+        The transformed initial_value after current_steps changes
     """
 
     def __init__(self, initial_value, end_value, adaptive_rate):
@@ -43,20 +75,20 @@ class InverseAdapter(BaseAdapter):
 
     Parameters
     ----------
-        initial_value : float,
-            Initial value to be adapted
-        end_value : float,
-            The final (asymptotic) value that the initial_value can take
-        adaptive_rate : float,
-            Controls how fast the initial_value approaches the end_value
+    initial_value : float,
+        Initial value to be adapted
+    end_value : float,
+        The final (asymptotic) value that the initial_value can take
+    adaptive_rate : float,
+        Controls how fast the initial_value approaches the end_value
 
     Attributes
     ----------
-        current_step : int,
-            The current number of iterations that the adapter has run
+    current_step : int,
+        The current number of iterations that the adapter has run
 
-        current_value : float,
-            The transformed initial_value after current_steps changes
+    current_value : float,
+        The transformed initial_value after current_steps changes
     """
 
     def __init__(self, initial_value, end_value, adaptive_rate):
@@ -78,20 +110,20 @@ class PotentialAdapter(BaseAdapter):
 
     Parameters
     ----------
-        initial_value : float,
-            Initial value to be adapted
-        end_value : float,
-            The final (asymptotic) value that the initial_value can take
-        adaptive_rate : float,
-            Controls how fast the initial_value approaches the end_value
+    initial_value : float,
+        Initial value to be adapted
+    end_value : float,
+        The final (asymptotic) value that the initial_value can take
+    adaptive_rate : float,
+        Controls how fast the initial_value approaches the end_value
 
     Attributes
     ----------
-        current_step : int,
-            The current number of iterations that the adapter has run
+    current_step : int,
+        The current number of iterations that the adapter has run
 
-        current_value : float,
-            The transformed initial_value after current_steps changes
+    current_value : float,
+        The transformed initial_value after current_steps changes
     """
 
     def __init__(self, initial_value, end_value, adaptive_rate):